+++ toolbox/tests/test_slugify.py
@@ -1,4 +1,4 @@
-from .. import slugify
+from ..core import slugify
 
 
 def test_slugify_english():