+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## master (unreleased)
 
-* Added a "language count" on the homepage (#43)
+* Added a "language count" on the homepage (#43).
+* Build a raw, but readable PDF for every language (#21).
 
 ## 2.3.0 (2016-09-16)